Changes variable access method from dot notation to getters in LICParameters
Since it is convetion (and best practice) to use getters and setters instead of dot notation to access
an object's instance variables the method of accessing variables was updated to fit the aformentioned
convetion.

The tests for LICParameter also had to be updated to correctly access the objects instance variables.