Creates LICParameter class (#36)
* Creates LICParameter class
The class stores all parameters related to the LICs. Unit tests are added to all attributes to make sure that one is able
to retrieve all parameters.

All .class files and the target-folder are added to .gitignore.

* Changes variable access method from dot notation to getters in LICParameters

Since it is convetion (and best practice) to use getters and setters instead of dot notation to access
an object's instance variables the method of accessing variables was updated to fit the aformentioned
convetion.

The tests for LICParameter also had to be updated to correctly access the objects instance variables.