Creates Vector class, fixes #40
Creates a Vector class. The vector consists of an underlying array with a generic type so that one is able to create vector objects of
different types. A vector object can be created with a specified length, or a length and initial vector values.

Setter and getters as well as a string formatting method are exposed.