Creates Vector class, feat #40 (#44)
* Creates Vector class, fixes #40

Creates a Vector class. The vector consists of an underlying array with a generic type so that one is able to create vector objects of
different types. A vector object can be created with a specified length, or a length and initial vector values.

Setter and getters as well as a string formatting method are exposed.

* Creates tests for Vector.java, fixes #40

Multiple unit tests for the Vector class. Tests the setters and getters as well as the string formatter.

* Added documentation to class

* Checks initial values for vector

Checks if the vector length and the length of the initial values are the same. If they are not, an error is throws.

* Tests constructor of vector

Tests if vector object instantiated with a length and initial values throws an error if their length is not the same.