Adds LIC2 and unit tests. Fixes #4
Adds the method to compute LIC2. The method loops over all the points in the input and determines if three consecutive points
have an angle > pi + epsilon or angle < pi - epsilon by using the law of cosines.

Adds unit tests that tests the two positive cases, and two negative cases (when angle > pi - epsilon and when the length of the point input
is < 3, because then 3 consecutive points do not exist).