Implements LIC2, fixes #4 (#64)
* Adds method and test to compare location of two points, used in solving #4

Adds hasSameLocation method to the Point class. The method checks if the instantiated Point object
has the same location (i.e. same x and y coordinate) as another point object.

Unit tests to check if the function returns false (not same location) and true (same location) are added.

* Adds method length() to the vector class that returns the length of the vector, used to solve #4

* Adds LIC2 and unit tests. Fixes #4

Adds the method to compute LIC2. The method loops over all the points in the input and determines if three consecutive points
have an angle > pi + epsilon or angle < pi - epsilon by using the law of cosines.

Adds unit tests that tests the two positive cases, and two negative cases (when angle > pi - epsilon and when the length of the point input
is < 3, because then 3 consecutive points do not exist).